requirements.txt
# Core dependencies
numpy>=1.21.0
pandas>=1.3.0
scipy>=1.7.0

# Database connectors
sqlalchemy>=1.4.0
psycopg2-binary>=2.9.0
pymysql>=1.0.0
pyodbc>=4.0.0

# Data processing and analysis
openpyxl>=3.0.0
xlrd>=2.0.0

# Machine Learning
scikit-learn>=1.0.0
xgboost>=1.5.0
lightgbm>=3.3.0

# Visualization
matplotlib>=3.4.0
seaborn>=0.11.0
plotly>=5.3.0

# Configuration
pyyaml>=5.4.0
python-dotenv>=0.19.0

# Utilities
python-dateutil>=2.8.0
pytz>=2021.3
tqdm>=4.62.0

# Testing
pytest>=6.2.0
pytest-cov>=3.0.0

# Development
black>=21.9b0
flake8>=4.0.0
isort>=5.9.0
